clean up of stats
